requirements.txt
transformers>=4.0.0
torch>=1.7.0
Pillow>=8.0.0